requirements.txt
streamlit==1.29.0
Pillow==10.0.0
plotly==5.18.0